use graphql-client utils
