+++ packages/admin-api-client/package.json
@@ -66,7 +66,7 @@
     "!node_modules"
   ],
   "dependencies": {
-    "@shopify/graphql-client": "^0.3.0"
+    "@shopify/graphql-client": "^0.5.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
