+++ packages/admin-api-client/src/constants.ts
@@ -3,3 +3,4 @@ export const DEFAULT_CONTENT_TYPE = "application/json";
 export const DEFAULT_CLIENT_VERSION = "ROLLUP_REPLACE_CLIENT_VERSION";
 export const ACCESS_TOKEN_HEADER = "X-Shopify-Access-Token";
 export const ERROR_PREFIX = "Admin API Client:";
+export const CLIENT = "Admin API Client";
