--- packages/admin-api-client/src/utilities/index.ts
@@ -1,2 +1,0 @@
-export * from "./api-versions";
-export * from "./validations";
